Tidying duplicated code
